cli: Fix output filename selection for >10 images
Images by default are placed in output/img_N.jpg where
N is a monotonically increased index.

There's a logic bug in the code, however, where it neglects
filenames that have more than one digit as an index.

This commit changes the logic to handle multi-digit filenames,
and also adds an optimization to avoid reparsing the filename
to obtain the index back out later, by reusing the results of
the initial regex search.